API

ContinuumMechanicsBase.CauchyStressTensor

ContinuumMechanicsBase.CauchyStressTensor

ContinuumMechanicsBase.I₁

ContinuumMechanicsBase.I₂

ContinuumMechanicsBase.I₃

ContinuumMechanicsBase.J

ContinuumMechanicsBase.SecondPiolaKirchoffStressTensor

ContinuumMechanicsBase.SecondPiolaKirchoffStressTensor

ContinuumMechanicsBase.StrainEnergyDensity

ContinuumMechanicsBase.StrainEnergyDensity

Hyperelastics.HyperelasticProblem

Hyperelastics.Kawabata1981

Hyperelastics.Treloar1944Uniaxial

Hyperelastics.parameter_bounds

Hyperelastics.parameters

Incompressible Models

Compressible Models

Helper Functions

ContinuumMechanicsBase.CauchyStressTensor — Method

CauchyStressTensor(ψ::AbstractHyperelasticModel, λ⃗::AbstractVector, p; adb=AD.ForwardDiffBackend())

Returns the Cauchy stress tensor for the hyperelastic model ψ with the principle stretches λ⃗ with parameters p.

Fields:

ψ: Hyperelastic model

λ⃗: Vector of principal stretches

p: Model parameters

adb: Differentiation backend from AbstractDifferentiation.jl

ContinuumMechanicsBase.CauchyStressTensor — Method

CauchyStressTensor(ψ::AbstractHyperelasticModel, F::AbstractMatrix, p; adb=AD.ForwardDiffBackend())

Returns the Cauchy stress tensor for the hyperelastic model ψ with the deformation gradient F with parameters p.

Fields:

ψ: Hyperelastic model

F: Deformation gradient tensor

p: Model parameters

adb: Differentiation backend from AbstractDifferentiation.jl

ContinuumMechanicsBase.I₁ — Method

First stretch invariant - Currently requires the addition of 5 times the machine precision to allow AD to work correctly

$I_1(\vec{\lambda}) = \lambda_1^2+\lambda_2^2+\lambda_3^2 + 5\varepsilon$

ContinuumMechanicsBase.I₂ — Method

Second Stretch invariant

$I_2(\vec{\lambda}) = \lambda_1^{-2}+\lambda_2^{-2}+\lambda_3^{-2}$

ContinuumMechanicsBase.I₃ — Method

Third Stretch invariant

$I_3(\vec{\lambda}) = (\lambda_1\lambda_2\lambda_3)^2$

ContinuumMechanicsBase.J — Method

Volumetric Stretch

$J(\vec{\lambda}) = \lambda_1\lambda_2\lambda_3$

ContinuumMechanicsBase.SecondPiolaKirchoffStressTensor — Method

SecondPiolaKirchoffStressTensor(ψ::AbstractHyperelasticModel, F::AbstractMatrix, p; adb=AD.ForwardDiffBackend())

Returns the second PK stress tensor for the hyperelastic model ψ with the deformation gradient F with parameters p.

Fields:

ψ: Hyperelastic model

F: Deformation gradient tensor

p: Model parameters

ContinuumMechanicsBase.SecondPiolaKirchoffStressTensor — Method

SecondPiolaKirchoffStressTensor(ψ::AbstractHyperelasticModel, λ⃗::AbstractVector, p; ad_type=AD.ForwardDiffBackend())

Returns the second PK stress tensor for the hyperelastic model ψ with the principle stretches λ⃗ with parameters p.

Fields:

ψ: Hyperelastic model

λ⃗: Vector of principal stretches

p: Model parameters

ContinuumMechanicsBase.StrainEnergyDensity — Method

StrainEnergyDensity(ψ, F, p)

Returns a function for the strain energy density function for the hyperelastic model based on calculating the principal stretches of the deformation gradient, F. The eigen values are found by the following procedure:

\[C = F^T \cdot F a = transpose(eigvecs(C)) C^\ast = (U^\ast)^2 = a^T \cdot C \cdot a \vec{\lambda} = diag(U)\]

Fields:

ψ: Hyperelastic model

F: Deformation gradient matrix

p: Model parameters

ContinuumMechanicsBase.StrainEnergyDensity — Method

StrainEnergyDensity(ψ, λ⃗, p)

Returns the strain energy density for the hyperelastic model ψ with the principle stretches λ⃗ with parameters p.

Fields:

ψ: Hyperelastic model

λ⃗: Vector of principal stretches

p: Model parameters

Hyperelastics.HyperelasticProblem — Function

HyperelasticProblem(ψ::AbstractHyperelasticModel, test::AbstractHyperelasticTest, u₀, ps=Nothing; adb=AD.ForwardDiffBackend(), loss=L2DistLoss(), adtype=Optimization.AutoForwardDiff())

HyperelasticProblem(ψ::AbstractHyperelasticModel, tests::Vector{<:AbstractHyperelasticTest}, u₀, ps=Nothing; adb=AD.ForwardDiffBackend(), loss=L2DistLoss(), adtype=Optimization.AutoForwardDiff())

Creates an OptimizationProblem for use in Optimization.jl to find the optimal parameters.

Fields:

ψ: material model to use

test or tests: A single or vector of hyperelastics tests to use when fitting the parameters

u₀: Initial guess for parameters

ps: Any additional parameters for calling predict

adb: Select differentiation type from ADTypes.jl. The type is automatically applied to the type of AD applied to the Optimization Problem also.

loss: Loss function from LossFunctions.jl

Hyperelastics.Kawabata1981 — Method

Kawabata1981(λ₁)

Biaxial Experimental Data from Kawabata et al. The model is more challenging to correctly fit a hyperelastic model to and is proposed as a better test than the Treloar1944 simple tension dataset.

Data provided for fixed λ₁ of:

1.040

1.060

1.080

1.100

1.120

1.14

1.16

1.2

1.24

1.3

1.6

1.9

2.2

2.5

2.8

3.1

3.4

3.7

Fields:

λ₁: Specification of λ₁ stretch for the data.

Kawabata S, Matsuda M, Tei K, Kawai H. Experimental survey of the strain energy density function of isoprene rubber vulcanizate. Macromolecules. 1981 Jan;14(1):154-62.

Hyperelastics.Treloar1944Uniaxial — Method

Treloar1944Uniaxial()

Uniaxial data for tension of 8% S Rubber at 20C from Fig 3 of Treloar 1944. This is commonly used for testing hyperelastic models.

Treloar LR. Stress-strain data for vulcanized rubber under various types of deformation. Rubber Chemistry and Technology. 1944 Dec;17(4):813-25.

Hyperelastics.parameter_bounds — Method

parameter_bounds(ψ::AbstractHyperelasticModel, test::AbstractHyperelasticTest) parameter_bounds(ψ::AbstractHyperelasticModel, tests::Vector{AbstractHyperelasticTest})

Returns a tuple of the parameter bounds provided the experimental data and model

Fields

ψ: Hyperelastic model

test or tests: The test or vector of tests to use in finding the parameter bounds.

Hyperelastics.parameters — Method

parameters(ψ::AbstractHyperelasticModel)

Returns a tuple of the parameters required for the model

Fields

ψ: Hyperelastics model