API Reference

Incompressible Models

Hyperelastics.ABGI — Method

ABGI(; ℒinv)

Model:

\[W = W_{Arruda-Boyce} + \frac{G_e}{n} \left(\sum_{i=1}^{3}\lambda_i^n-3\right)\]

Arguments:

ℒinv = TreloarApproximation(): Sets the inverse Langevin approxamationused (default = TreloarApproximation())

Parameters:

μ

N

Ge

n

Meissner B, Matějka L. A Langevin-elasticity-theory-based constitutiveequation for rubberlike networks and its comparison with biaxialstress–strain data. Part I. Polymer. 2003 Jul 1;44(16):4599-610.

Hyperelastics.AffineMicroSphere — Method

AffineMicroSphere(; ℒinv, n)

Model:

See Paper

Arguments:

ℒinv=TreloarApproximation(): Sets the inverse Langevin approxamation used

n::Int = 21: Number of quadrature points for the spherical integration

Parameters:

μ

N

Miehe C, Göktepe S, Lulei F. A micro-macro approach to rubber-like materials—part I: the non-affine micro-sphere model of rubber elasticity. Journal of the Mechanics and Physics of Solids. 2004 Nov 1;52(11):2617-60.

Hyperelastics.Alexander — Method

Alexander()

Model:

\[W = \frac{C_1 \sqrt{\pi}\text{erfi}\big(\sqrt{k}(I_1-3)\big)}{2\sqrt{k}}+C_2\log{\frac{I_2-3+\gamma}{\gamma}}+C_3(I_2-3)\]

Parameters:

μ

C₁

C₂

C₃

k

γ

Alexander H. A constitutive relation for rubber-like materials. International Journal of Engineering Science. 1968 Sep 1;6(9):549-63.

Hyperelastics.Amin — Method

Amin()
Amin(type)

Model:

\[W = C_1 (I_1 - 3) + \frac{C_2}{N + 1} (I_1 - 3)^{N + 1} + \frac{C_3}{M + 1} (I_1 - 3)^{M + 1} + C_4 (I_2 - 3)\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

C1

C2

C3

C4

N

M

Amin AF, Wiraguna SI, Bhuiyan AR, Okui Y. Hyperelasticity model for finite element analysis of natural and high damping rubbers in compression and shear. Journal of engineering mechanics. 2006 Jan;132(1):54-64.

Hyperelastics.AnsarriBenam — Method

AnsarriBenam(; ...)
AnsarriBenam(type; n)

Model:

\[W = \frac{3(n-1)}{2n}\mu N \left[\frac{1}{3N(n-1)}(I_1 - 3) - \log{\frac{I_1 - 3N}{3 -3N}} \right]\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

ℒinv=TreloarApproximation(): Sets the inverse Langevin approxamation used (default = ``)

n::Int=3: Sets the order of the model

Parameters:

μ

n

N

Anssari-Benam A. On a new class of non-Gaussian molecular-based constitutive models with limiting chain extensibility for incompressible rubber-like materials. Mathematics and Mechanics of Solids. 2021 Nov;26(11):1660-74.

Hyperelastics.ArmanNarooei — Method

ArmanNarooei()

Model:

\[W = \sum\limits_{i=1}^{N} A_i\big[\exp{m_i(\lambda_1^{\alpha_i}+\lambda_2^{\alpha_i}+\lambda_3^{\alpha_i}-3)}-1] + B_i\big[\exp{n_i(\lambda_1^{-\beta_i}+\lambda_2^{-\beta_i}+\lambda_3^{-\beta_i}-3)}-1]\]

Parameters:

A⃗

B⃗

m⃗

n⃗

α⃗

β⃗

Narooei K, Arman M. Modification of exponential based hyperelastic strain energy to consider free stress initial configuration and Constitutive modeling. Journal of Computational Applied Mechanics. 2018 Jun 1;49(1):189-96.

Hyperelastics.ArrudaBoyce — Method

ArrudaBoyce(; ...)
ArrudaBoyce(type; ℒinv)

Model:

\[W = \mu N \left( \frac{\lambda_{chain}}{\sqrt{N}} \beta + \log\left(\frac{\beta}{\sinh\beta}\right) \right)\]

where

\[\beta = \mathcal{L}^{-1}\left(\frac{\lambda_{chain}}{\sqrt{N}}\right)\]

and

\[\lambda_{chain} = \sqrt{\frac{I_1}{3}}\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or InvariantForm()

ℒinv=TreloarApproximation(): Sets the inverse Langevin approxamation used

Parameters:

μ: Small strain shear modulus

N: Square of the locking stretch of the network.

Arruda EM, Boyce MC. A three-dimensional constitutive model for the large stretch behavior of rubber elastic materials. Journal of the Mechanics and Physics of Solids. 1993 Feb 1;41(2):389-412.

Hyperelastics.Attard — Method

Attard()

Model:

\[W = \sum\limits_{i=1}^N\frac{A_i}{2i}(\lambda_1^{2i}+\lambda_2^{2i}+\lambda_3^{2i}-3) + \frac{B_i}{2i}(\lambda_1^{-2i}+\lambda_2^{-2i}+\lambda_3^{-2i}-3)\]

Parameters:

A⃗

B⃗

Attard MM, Hunt GW. Hyperelastic constitutive modeling under finite strain. International Journal of Solids and Structures. 2004 Sep 1;41(18-19):5327-50.

Hyperelastics.BahremanDarijani — Method

BahremanDarijani()
BahremanDarijani(type)

Model:

\[W = \sum\limits_{i = 1}{3}\sum\limits_{j=0}^{N} A_j (\lambda_i^{m_j}-1) + B_j(\lambda_i^{-n_j}-1)\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or InvariantForm()

Parameters:

A2

B2

A4

A6

Bahreman M, Darijani H. New polynomial strain energy function; application to rubbery circular cylinders under finite extension and torsion. Journal of Applied Polymer Science. 2015 Apr 5;132(13).

Hyperelastics.Beatty — Method

Beatty()
Beatty(type)

Model:

\[W = -\frac{G_0 I_m(I_m-3)}{2(2I_m-3)}\log\bigg(\frac{1-\frac{I_1-3}{I_m-3}}{1+\frac{I_1-3}{I_m}} \bigg)\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

G₀

Iₘ

Beatty MF. On constitutive models for limited elastic, molecular based materials. Mathematics and mechanics of solids. 2008 Jul;13(5):375-87.

Hyperelastics.Bechir4Term — Method

Bechir4Term()

Model:

\[W = C_1^1(I_1-3)+\sum\limits_{n=1}^{2}\sum\limits_{r=1}^{2}C_n^{r}(\lambda_1^{2n}+\lambda_2^{2n}+\lambda_3^{2n}-3)^r\]

Parameters:

C11

C12

C21

C22

Khajehsaeid H, Arghavani J, Naghdabadi R. A hyperelastic constitutive model for rubber-like materials. European Journal of Mechanics-A/Solids. 2013 Mar 1;38:144-51.

Hyperelastics.BechirChevalier — Method

BechirChevalier(; ℒinv)

Model:

\[W = W_{3Chain}(\mu_f, N_3)+W_{8Chain}(\frac{\mu_c}{3}, N_8)\]

where:

\[\mu_f = \rho\sqrt{\frac{I_1}{3N_8}}\]

\[\mu_c = \bigg(1-\frac{\eta\alpha}{\sqrt{N_3}}\bigg)\mu_0\]

\[\alpha = \max{\lambda_1, \lambda_2, \lambda_3}\]

Arguments:

ℒinv=TreloarApproximation(): Sets the inverse Langevin approxamation used

Parameters:

μ₀

η

ρ

N₃

N₈

Bechir H, Chevalier L, Idjeri M. A three-dimensional network model for rubber elasticity: The effect of local entanglements constraints. International journal of engineering science. 2010 Mar 1;48(3):265-74.

Hyperelastics.Beda — Method

Beda()
Beda(type)

Model:

\[W = \frac{C_1}{\alpha}(I_1-3)^{\alpha}+C_2(I_1-3)+\frac{C_3}{\zeta}(I_1-3)^{\zeta}+\frac{K_1}{\beta}(I_2-3)^\beta\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

C1

C2

C3

K1

α

β

ζ

Beda T. Reconciling the fundamental phenomenological expression of the strain energy of rubber with established experimental facts. Journal of Polymer Science Part B: Polymer Physics. 2005 Jan 15;43(2):125-34.

Hyperelastics.Biderman — Method

Biderman()
Biderman(type)

Model:

\[W = \sum\limits_{i,j=0}^{3, 1}C_{i,j}(I_1-3)^i(I_2-3)^j\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or InvariantForm()

Parameters:

C10

C01

C20

C30

Biderman VL. Calculation of rubber parts. Rascheti na prochnost. 1958;40.

Hyperelastics.Bootstrapped8Chain — Method

Bootstrapped8Chain(; ℒinv)

Model:

\[W = W_8(\frac{\sum\lambda}{\sqrt{3N}}-\frac{\lambda_{chain}}{\sqrt{N}})+W_{8}(\frac{\lambda_{chain}}{\sqrt{N}})\]

where:

\[W_8(x) = \mu N (x \mathcal{L}^{-1}(x) + \log\frac{\mathcal{L}^{-1}(x)}{\sinh\mathcal{L}^{-1}(x)})\]

and

\[\lambda_{chain} = \sqrt{\frac{I_1}{3}}\]

Arguments:

ℒinv=TreloarApproximation(): Sets the inverse Langevin approximation used.

Parameters:

μ

N

Miroshnychenko D, Green WA, Turner DM. Composite and filament models for the mechanical behaviour of elastomeric materials. Journal of the Mechanics and Physics of Solids. 2005 Apr 1;53(4):748-70. Miroshnychenko D, Green WA. Heuristic search for a predictive strain-energy function in nonlinear elasticity. International Journal of Solids and Structures. 2009 Jan 15;46(2):271-86.

Hyperelastics.Carroll — Method

Carroll()
Carroll(type)

Model:

\[W = AI_1+BI_1^4+C\sqrt{I_2}\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or InvariantForm()

Parameters:

A

B

C

Carroll M. A strain energy function for vulcanized rubbers. Journal of Elasticity. 2011 Apr;103(2):173-87.

Hyperelastics.ChevalierMarco — Method

ChevalierMarco()
ChevalierMarco()

Model:

\[W = \int\limits_{3}^{I_1(\vec\lambda)} \exp\bigg(\sum\limits_{i=0}^{N}a_i(I_1-3)^i\bigg)\text{d}I_1+ \int\limits_{3}^{I_2(\vec\lambda)} \sum\limits_{i=0}^{n}\frac{b_i}{I_2^i}\text{d}I_2\]

\[[\mathbf{S}] = 2(I-\frac{\partial W}{\partial I_1} - C^{-2}\frac{\partial W}{\partial I_2})\]

\[[\mathbf{\sigma}] = \mathbf{F} \cdot \mathbf{S}\]

Parameters:

a⃗

b⃗

Note:

Model is not compatible with AD. A method for accessing the Second Piola Kirchoff Tensor and Cauchy Stress Tensor have been implemented.

Chevalier L, Marco Y. Tools for multiaxial validation of behavior laws chosen for modeling hyper‐elasticity of rubber‐like materials. Polymer Engineering & Science. 2002 Feb;42(2):280-98.

Hyperelastics.ConstrainedJunction — Method

ConstrainedJunction()

Model:

\[W = G_c (I_1-3)+ \frac{\nu k T}{2}\left(\sum\limits_{i=1}^{3}\kappa\frac{\lambda_i-1}{\lambda_i^2+\kappa}+\log{\frac{\lambda_i^2+\kappa}{1+\kappa}}-\log{\lambda_i^2}\right)\]

Parameters:

Gc

νkT

κ

Flory PJ, Erman B. Theory of elasticity of polymer networks. 3. Macromolecules. 1982 May;15(3):800-6. Erman B, Flory PJ. Relationships between stress, strain, and molecular constitution of polymer networks. Comparison of theory with experiments. Macromolecules. 1982 May;15(3):806-11.

Hyperelastics.ContinuumHybrid — Method

ContinuumHybrid()

Model:

\[W = K_1(I_1-3)+K_2\log\frac{I_2}{3}+\frac{\mu}{\alpha}(\lambda_1^\alpha+\lambda_2^\alpha+\lambda^\alpha-3)\]

Parameters:

K₁

K₂

α

μ

Beda T, Chevalier Y. Hybrid continuum model for large elastic deformation of rubber. Journal of applied physics. 2003 Aug 15;94(4):2701-6.

Hyperelastics.DavidsonGoulbourne — Method

DavidsonGoulbourne()

Model:

\[W = \frac{G_c I_1}{6}-G_c\lambda_{max}\log\left(3\lambda_{max}^2-I_1\right)+G_e\sum\limits_{i=1}^{3}\left(\lambda_i+\frac{1}{\lambda_i}\right)\]

Parameters:

Gc

Ge

λmax

Davidson JD, Goulbourne NC. A nonaffine network model for elastomers undergoing finite deformations. Journal of the Mechanics and Physics of Solids. 2013 Aug 1;61(8):1784-97.

Hyperelastics.DavisDeThomas — Method

DavisDeThomas()
DavisDeThomas(type)

Model:

\[W = \frac{A}{2(1-\frac{n}{2})}(I_1-3+C^2)^{1-\frac{n}{2}}+k(I_1-3)^2\]

Arguments

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

A

n

C

k

Davies CK, De DK, Thomas AG. Characterization of the behavior of rubber for engineering design purposes. 1. Stress-strain relations. Rubber chemistry and technology. 1994 Sep;67(4):716-28.

Hyperelastics.EdwardVilgis — Method

EdwardVilgis()

Model:

\[W = \frac{1}{2}N_C\Bigg[\frac{(1-\alpha^2)I_1}{1-\alpha^2I_1}+\log(1-\alpha^2I_1)\Bigg]+\frac{1}{2}N_S\Bigg[\sum_{i=1}^{3}\Big\{\frac{(1+\eta)(1-\alpha^2)\lambda_i^2}{( 1+\eta\lambda_i^2)(1-\alpha^2I_1)}+\log(1+\eta\lambda_i^2)\Big\}+\log(1-\alpha^2I_1)\Bigg]\]

Parameters:

Ns: Number of sliplinks

Nc: Number of crosslinks

α: A measure of chain inextensibility

η: A measure of the amount of chain slippage

Note:

Since α and η result from the same mechanism, they should be of approximately the same order of magnitude. Large differences between the two may indicate an issue with the optimizer or initial guess.

Edwards SF, Vilgis T. The effect of entanglements in rubber elasticity. Polymer. 1986 Apr 1;27(4):483-92.

Hyperelastics.ExpLn — Method

ExpLn()
ExpLn(type)

Model:

\[W = A\bigg[\frac{1}{a}\exp{(a(I_1-3))}+b(I_1-2)(1-\log{I_1-2})-\frac{1}{a}-b\bigg]\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

A

a

b

Khajehsaeid H, Arghavani J, Naghdabadi R. A hyperelastic constitutive model for rubber-like materials. European Journal of Mechanics-A/Solids. 2013 Mar 1;38:144-51.

Hyperelastics.ExtendedTubeModel — Method

ExtendedTubeModel()

Model:

\[W = \frac{G_c}{2}\bigg[\frac{(1-\delta^2)(I_1-3)}{1-\delta^2(I_1-3)}+\log{(1-\delta^2(I_1-3))}\bigg]+\frac{2G_e}{\beta^2}\sum\limits_{i=1}^{3}(\lambda_i^{-\beta}-1)\]

Parameters:

Gc

Ge

δ

β

Kaliske M, Heinrich G. An extended tube-model for rubber elasticity: statistical-mechanical theory and finite element implementation. Rubber Chemistry and Technology. 1999 Sep;72(4):602-32.

Hyperelastics.FullNetwork — Method

FullNetwork(; ℒinv)

Model:

\[W = (1-\rho)W_{3Chain}+\rho W_{8chain}\]

Arguments:

ℒinv=TreloarApproximation(): Sets the inverse Langevin approxamation used

Parameters:

μ

N

ρ

Treloar LR, Riding G. A non-Gaussian theory for rubber in biaxial strain. I. Mechanical properties. Proceedings of the Royal Society of London. A. Mathematical and Physical Sciences. 1979 Dec 31;369(1737):261-80. Wu PD, van der Giessen E. On improved 3-D non-Gaussian network models for rubber elasticity. Mechanics research communications. 1992 Sep 1;19(5):427-33. Wu PD, Van Der Giessen E. On improved network models for rubber elasticity and their applications to orientation hardening in glassy polymers. Journal of the Mechanics and Physics of Solids. 1993 Mar 1;41(3):427-56.

Hyperelastics.FungDemiray — Method

FungDemiray()
FungDemiray(type)

Model:

\[W = \frac{\mu}{2 * b} (\exp(b(I_1 - 3)) - 1)\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

μ

b

Fung YC. Elasticity of soft tissues in simple elongation. American Journal of Physiology-Legacy Content. 1967 Dec 1;213(6):1532-44. Demiray H. A note on the elasticity of soft biological tissues. Journal of biomechanics. 1972 May 1;5(3):309-11.

Hyperelastics.GenYeoh — Method

GenYeoh()
GenYeoh(type)

Model:

\[W = K_1 (I_1 - 3)^m + K_2 * (I_1 - 3)^p + K_3 * (I_1 - 3)^q\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

K1

K2

K3

m

p

q

Hohenberger TW, Windslow RJ, Pugno NM, Busfield JJ. A constitutive model for both low and high strain nonlinearities in highly filled elastomers and implementation with user-defined material subroutines in ABAQUS. Rubber Chemistry and Technology. 2019;92(4):653-86.

Hyperelastics.GeneralConstitutiveModel — Method

GeneralConstitutiveModel()

Model:

\[W = G_c N \log\bigg(\frac{3N+\frac{1}{2}I_1}{3N-I_1}\bigg)+G_e\sum\limits_{i=1}^{3}\frac{1}{\lambda_I}\]

Parameters:

Gc

Ge

N

Xiang Y, Zhong D, Wang P, Mao G, Yu H, Qu S. A general constitutive model of soft elastomers. Journal of the Mechanics and Physics of Solids. 2018 Aug 1;117:110-22.

Hyperelastics.Gent — Method

Gent()
Gent(type)

Model:

\[W = -\frac{\mu J_m}{2}\log{\bigg(1-\frac{I_1-3}{J_m}\bigg)}\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

μ: Small strain shear modulus

Jₘ: Limiting stretch invariant

Gent AN. A new constitutive relation for rubber. Rubber chemistry and technology. 1996 Mar;69(1):59-61.

Hyperelastics.GentThomas — Method

GentThomas()
GentThomas(type)

Model:

\[W = C_1(I_1-3)+C_2\log(\frac{I_2}{3})\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Paramters:

C1

C2

Gent AN, Thomas AG. Forms for the stored (strain) energy function for vulcanized rubber. Journal of Polymer Science. 1958 Apr;28(118):625-8.

Hyperelastics.GornetDesmorat — Method

GornetDesmorat()
GornetDesmorat(type)

Model:

\[W = h_1\int\exp{h_3(I_1-3)^2}\text{d}I_1+3h_2\int\frac{1}{\sqrt{I_2}}\text{d}I_2 = \frac{h_1 \sqrt{\pi} \text{erfi}(\sqrt{h_3}(I_1-3)^2)}{2\sqrt{h_3}}+6h_2\sqrt{I_2}\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

h₁

h₂

h₃

Notes:

The differential form was original form and the closed form SEF was determine via symbolic integration in Mathematica. The model is not compatible with AD and has methods for the Second Piola Kirchoff Stress Tensor and Cauchy Stress Tensor implemented.

Gornet L, Marckmann G, Desmorat R, Charrier P. A new isotropic hyperelastic strain energy function in terms of invariants and its derivation into a pseudo-elastic model for Mullins effect: application to finite element analysis. Constitutive Models for Rubbers VII. 2012:265-71.

Hyperelastics.Gregory — Method

Gregory()
Gregory(type)

Model:

\[W = \frac{A}{2-n}(I_1-3+C^2)^{1-\frac{n}{2}}+\frac{B}{2+m}(I_1-3+C^2)^{1+\frac{m}{2}}\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

A

B

C

m

n

Gregory IH, Muhr AH, Stephens IJ. Engineering applications of rubber in simple extension. Plastics rubber and composites processing and applications. 1997;26(3):118-22.

Hyperelastics.HainesWilson — Method

HainesWilson()
HainesWilson(type)

Model:

\[W = \sum\limits_{i,j=0}^{3, 2}C_{i,j}(I_1-3)^i(I_2-3)^j\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or InvariantForm()

Parameters:

C10

C01

C11

C02

C20

C30

Haines DW, Wilson WD. Strain-energy density function for rubberlike materials. Journal of the Mechanics and Physics of Solids. 1979 Aug 1;27(4):345-60.

Hyperelastics.HartSmith — Method

HartSmith()
HartSmith(type)

Model:

\[W = \frac{G\exp{(-9k_1+k_1I_1)}}{k_1}+Gk_2\log{I_2}\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

G

k₁

k₂

Hart-Smith LJ. Elasticity parameters for finite deformations of rubber-like materials. Zeitschrift für angewandte Mathematik und Physik ZAMP. 1966 Sep;17(5):608-26.

Hyperelastics.HartmannNeff — Method

HartmannNeff()
HartmannNeff(type)

Model:

\[W = \sum\limits_{i,j=0}^{M,N}C_{i,0}(I_1-3)^i -3\sqrt{3}^j+\alpha(I_1-3)\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or InvariantForm()

Parameters:

α

Ci⃗0

C0j⃗

Hartmann S, Neff P. Polyconvexity of generalized polynomial-type hyperelastic strain energy functions for near-incompressibility. International journal of solids and structures. 2003 Jun 1;40(11):2767-91.

Hyperelastics.HauptSedlan — Method

HauptSedlan()
HauptSedlan(type)

Model:

\[W = \sum\limits_{i,j=0}^{3, 2}C_{i,j}(I_1-3)^i(I_2-3)^j\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or InvariantForm()

Parameters:

C10

C01

C11

C02

C30

Haupt P, Sedlan K. Viscoplasticity of elastomeric materials: experimental facts and constitutive modelling. Archive of Applied Mechanics. 2001 Mar;71(2):89-109.

Hyperelastics.HorganMurphy — Method

HorganMurphy()

Model:

\[W = -\frac{2\mu J_m}{c^2}\log\left(1-\frac{\lambda_1^c+\lambda_2^c+\lambda_3^c-3}{J_m}\right)\]

Parameters:

μ

Jₘ

c

Horgan CO, Murphy JG. Limiting chain extensibility constitutive models of Valanis–Landel type. Journal of Elasticity. 2007 Feb;86(2):101-11.

Hyperelastics.HorganSaccomandi — Method

HorganSaccomandi()
HorganSaccomandi(type)

Model:

\[W = -\frac{\mu J}{2}\log\bigg(\frac{J^3-J^2I_1+JI_2-1}{(J-1)^3}\bigg)\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

μ

J

Horgan CO, Saccomandi G. Constitutive models for compressible nonlinearly elastic materials with limiting chain extensibility. Journal of Elasticity. 2004 Nov;77(2):123-38.> Horgan CO, Saccomandi G. Constitutive models for atactic elastomers. InWaves And Stability In Continuous Media 2004 (pp. 281-294).

Hyperelastics.HossMarczakI — Method

HossMarczakI()
HossMarczakI(type)

Model:

\[W = \frac{\alpha}{\beta}(1-\exp{-\beta(I_1-3)})+\frac{\mu}{2b}\bigg((1+\frac{b}{n}(I_1-3))^n -1\bigg)\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

α

β

μ

b

n

Note:

The authors suggested this model for low strains.

Hoss L, Marczak RJ. A new constitutive model for rubber-like materials. Mecánica Computacional. 2010;29(28):2759-73.

Hyperelastics.HossMarczakII — Method

HossMarczakII()
HossMarczakII(type)

Model:

\[W = \frac{\alpha}{\beta}(1-\exp{-\beta(I_1-3)})+\frac{\mu}{2b}\bigg((1+\frac{b}{n}(I_1-3))^n -1\bigg)+C_2\log(\frac{I_2}{3})\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

α

β

μ

b

n

C2

Note:

The authors suggests this model for high strains.

Hoss L, Marczak RJ. A new constitutive model for rubber-like materials. Mecánica Computacional. 2010;29(28):2759-73.

Hyperelastics.Isihara — Method

Isihara()
Isihara(type)

Model:

\[W = \sum\limits_{i,j=0}^{2, 1}C_{i,j}(I_1-3)^i(I_2-3)^j\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or InvariantForm()

Parameters:

C10

C20

C01

Isihara A, Hashitsume N, Tatibana M. Statistical theory of rubber‐like elasticity. IV.(two‐dimensional stretching). The Journal of Chemical Physics. 1951 Dec;19(12):1508-12.

Hyperelastics.JamesGreenSimpson — Method

JamesGreenSimpson()
JamesGreenSimpson(type)

Model:

\[W = \sum\limits_{i,j=0}^{3, 1}C_{i,j}(I_1-3)^i(I_2-3)^j\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or InvariantForm()

Parameters:

C10

C01

C11

C20

C30

James AG, Green A, Simpson GM. Strain energy functions of rubber. I. Characterization of gum vulcanizates. Journal of Applied Polymer Science. 1975 Jul;19(7):2033-58.

Hyperelastics.KhiemItskov — Method

KhiemItskov()
KhiemItskov(type)

Model:

\[W = \mu_c \kappa n \log\bigg(\frac{\sin(\frac{\pi}{\sqrt{n}})(\frac{I_1}{3})^{\frac{q}{2}}}{\sin(\frac{\pi}{\sqrt{n}}(\frac{I_1}{3})^{\frac{q}{2}}}\bigg)+\mu_t\big[\frac{I_2}{3}^{1/2} - 1 \big]\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

μcκ

n

q

μt

Khiêm VN, Itskov M. Analytical network-averaging of the tube model:: Rubber elasticity. Journal of the Mechanics and Physics of Solids. 2016 Oct 1;95:254-69.

Hyperelastics.Knowles — Method

Knowles()
Knowles(type)

Model:

\[W = \frac{\mu}{2b}\left(\left(1+\frac{b}{n}(I_1-3)\right)^n-1\right)\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

μ

b

n

Knowles JK. The finite anti-plane shear field near the tip of a crack for a class of incompressible elastic solids. International Journal of Fracture. 1977 Oct;13(5):611-39.

Hyperelastics.LambertDianiRey — Method

LambertDianiRey()
LambertDianiRey(type)

Model:

\[W = \int\limits_{3}^{I_1}\exp\bigg(\sum\limits_{i=0}^{n}a_i(I_1-3)^i\bigg)\text{d}I_1+\int\limits_{3}^{I_2}\sum\limits_{j=0}^{m}b_i\log(I_2)^i\text{d}I_2\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

a⃗

b⃗

Lambert-Diani J, Rey C. New phenomenological behavior laws for rubbers and thermoplastic elastomers. European Journal of Mechanics-A/Solids. 1999 Nov 1;18(6):1027-43.

Hyperelastics.Lim — Method

Lim(; ...)
Lim(type; ℒinv)

Model:

\[W = \left(1-f\left(\frac{I_1-3}{\hat{I_1}-3}\right)\right)W_{NeoHookean}(μ₁)+fW_{ArrudaBoyce}(μ₂, N)\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

ℒinv=TreloarApproximation(): Sets the inverse Langevin approxamation used

Parameters:

μ₁

μ₂

N

Î₁

Lim GT. Scratch behavior of polymers. Texas A&M University; 2005.

Hyperelastics.Lion — Method

Lion()
Lion(type)

Model:

\[W = \sum\limits_{i,j=0}^{5,1}C_{i,j}(I_1-3)^i(I_2-3)^j\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or InvariantForm()

Parameters:

C10

C01

C50

Lion A. On the large deformation behaviour of reinforced rubber at different temperatures. Journal of the Mechanics and Physics of Solids. 1997 Nov 1;45(11-12):1805-34.

Hyperelastics.LopezPamies — Method

LopezPamies()
LopezPamies(type)

Model:

\[W = \frac{3^{1 - \alpha_i}}{2\alpha_i} \mu_i (I_1^{\alpha_i} - 3^{\alpha_i})\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

α⃗

μ⃗

Lopez-Pamies O. A new I1-based hyperelastic model for rubber elastic materials. Comptes Rendus Mecanique. 2010 Jan 1;338(1):3-11.

Hyperelastics.MCC — Method

MCC()

Model:

\[W = \frac{1}{2}\zeta k T \sum\limits_{i=1}^{3}(\lambda_i^2-1)+\frac{1}{2}\mu k T\sum\limits_{i=1}^{3}[B_i+D_i-\log{(1+B_i)}-\log{(1+D_i)}]\]

where:

\[B_i = \frac{\kappa^2(\lambda_i^2-1)}{(\lambda_i^2+\kappa)^2}\]

and

\[D_i = \frac{\lambda_i^2 B_i}{\kappa}\]

Parameters:

ζkT

μkT

κ

Erman B, Monnerie L. Theory of elasticity of amorphous networks: effect of constraints along chains. Macromolecules. 1989 Aug;22(8):3342-8.

Hyperelastics.MansouriDarijani — Method

MansouriDarijani()
MansouriDarijani(type)

Model:

\[W = A_1\exp{m_1(I_1-3)-1}+B_1\exp{n_1(I_2-3)-1}\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

A1

m1

B1

n1

Mansouri MR, Darijani H. Constitutive modeling of isotropic hyperelastic materials in an exponential framework using a self-contained approach. International Journal of Solids and Structures. 2014 Dec 1;51(25-26):4316-26.

Hyperelastics.ModifiedFloryErman — Method

ModifiedFloryErman(; ℒinv)

Model:

\[W = W_{\text{Arruda-Boyce}}+\sum\limits_{i=1}^{3}\frac{\mu}{2}[B_i+D_i]\]

Arguments:

ℒinv=TreloarApproximation(): Sets the inverse Langevin approxamation used

Parameters:

μ

N

κ

Edwards SF. The statistical mechanics of polymerized material. Proceedings of the Physical Society (1958-1967). 1967 Sep 1;92(1):9.

Hyperelastics.ModifiedGregory — Method

ModifiedGregory()
ModifiedGregory(type)

Model:

\[W = \frac{A}{1+\alpha}(I_1-3+M^2)^{1+\alpha}+\frac{B}{1+\beta}(I_1-3+N^2)^{1+\beta}\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

A

α

M

B

β

N

He H, Zhang Q, Zhang Y, Chen J, Zhang L, Li F. A comparative study of 85 hyperelastic constitutive models for both unfilled rubber and highly filled rubber nanocomposite material. Nano Materials Science. 2021 Jul 16.

Hyperelastics.ModifiedYeoh — Method

ModifiedYeoh()
ModifiedYeoh(type)

Model:

\[W = C_{10} * (I_1 - 3) + C_{20} * (I_1 - 3)^2 + C_{30} * (I_1 - 3)^3 + \alpha / \beta * (1 - \exp{-\beta * (I_1 - 3)})\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

C10

C20

C30

α

β

He H, Zhang Q, Zhang Y, Chen J, Zhang L, Li F. A comparative study of 85 hyperelastic constitutive models for both unfilled rubber and highly filled rubber nanocomposite material. Nano Materials Science. 2021 Jul 16.

Hyperelastics.MooneyRivlin — Method

MooneyRivlin()
MooneyRivlin(type)

Model:

\[W = C_{10}(I_1-3)+C_{01}(I_2-3)\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or InvariantForm()

Parameters:

C01

C10

Mooney M. A theory of large elastic deformation. Journal of applied physics. 1940 Sep;11(9):582-92.

Hyperelastics.NeoHookean — Method

NeoHookean()
NeoHookean(type)

Model:

\[W = \frac{\mu}{2}(I_1-3)\]

Arguments

type = PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

μ: Small strain shear modulus

Treloar LR. The elasticity of a network of long-chain molecules—II. Transactions of the Faraday Society. 1943;39:241-6.

Hyperelastics.NonaffineMicroSphere — Method

NonaffineMicroSphere(; ℒinv, n)

Model: See Paper

Arguments:

ℒinv=CohenRounded3_2(): Sets the inverse Langevin approximation used.

n=21: Order of quadrature for spherical integration

Parameters:

μ: Small strain shear modulus

N: Number of chain segments

p: Non-affine stretch parameter

U: Tube geometry parameter

q: Non-affine tube parameter

Miehe C, Göktepe S, Lulei F. A micro-macro approach to rubber-like materials—part I: the non-affine micro-sphere model of rubber elasticity. Journal of the Mechanics and Physics of Solids. 2004 Nov 1;52(11):2617-60.

Hyperelastics.NonaffineTube — Method

NonaffineTube()

Model:

\[W = G_c \sum\limits_{i=1}^{3}\frac{\lambda_i^2}{2}+G_e\sum\limits_{i=1}^{3}\lambda_i+\frac{1}{\lambda_i}\]

Parameters:

Gc

Ge

Rubinstein M, Panyukov S. Nonaffine deformation and elasticity of polymer networks. Macromolecules. 1997 Dec 15;30(25):8036-44.

Hyperelastics.Ogden — Method

Ogden()

Model:

\[W = \sum\limits_{i=1}^{N}\frac{\mu_i}{\alpha_i}(\lambda_1^{\alpha_i}+\lambda_2^{\alpha_i}+\lambda_3^{\alpha_i}-3)\]

Parameters:

μ⃗

α⃗

Ogden RW. Large deformation isotropic elasticity–on the correlation of theory and experiment for incompressible rubberlike solids. Proceedings of the Royal Society of London. A. Mathematical and Physical Sciences. 1972 Feb 1;326(1567):565-84.

Hyperelastics.PengLandel — Method

PengLandel()

Model:

\[W = E\sum\limits_{i=1}^{3}\bigg[\lambda_i - 1 - \log(\lambda_i) - \frac{1}{6}\log(\lambda_i)^2 + \frac{1}{18}\log(\lambda_i)^3-\frac{1}{216}\log(\lambda_i)^4\bigg]\]

Parameters:

E

Peng TJ, Landel RF. Stored energy function of rubberlike materials derived from simple tensile data. Journal of Applied Physics. 1972 Jul;43(7):3064-7.

Hyperelastics.PucciSaccomandi — Method

PucciSaccomandi()
PucciSaccomandi(type)

Model:

\[W = K\log{\frac{I_2}{3}}-\frac{\mu J_m}{2}\log{1-\frac{I_1-3}{J-m}}\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

K

μ

Jₘ

Pucci E, Saccomandi G. A note on the Gent model for rubber-like materials. Rubber chemistry and technology. 2002 Nov;75(5):839-52.

Hyperelastics.Shariff — Method

Shariff()

Model:

\[W = E\sum\limits_{i=1}^3\sum\limits_{j=1}^{N}|\alpha_j| \Phi_j(\lambda_i)\]

Parameters:

E

α⃗

Shariff MH. Strain energy function for filled and unfilled rubberlike material. Rubber chemistry and technology. 2000 Mar;73(1):1-8.

Hyperelastics.Swanson — Method

Swanson()
Swanson(type)

Model:

\[W = \sum\limits_{i=1}^{N} \frac{3}{2}(\frac{A_i}{1+\alpha_i}(\frac{I_1}{3})^{1+\alpha_i}+\frac{B_i}{1+\beta_i}(\frac{I_2}{3})^{1+\beta_i}\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

A⃗

α⃗

B⃗

β⃗

Swanson SR. A constitutive model for high elongation elastic materials.

Hyperelastics.TakamizawaHayashi — Method

TakamizawaHayashi()
TakamizawaHayashi(type)

Model:

\[W = -c\log\left1-\left(\frac{I_1-3}{J_m}\right)^2\right\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

c

Jₘ

Takamizawa K, Hayashi K. Strain energy density function and uniform strain hypothesis for arterial mechanics. Journal of biomechanics. 1987 Jan 1;20(1):7-17.

Hyperelastics.ThreeChainModel — Method

ThreeChainModel(; ℒinv)

Model:

\[W = \frac{\mu\sqrt{N}}{3}\sum\limits_{i=1}^{3}\bigg(\lambda_i\beta_i+\sqrt{N}\log\bigg(\frac{\beta_i}{\sinh \beta_i}\bigg)\bigg)\]

Arguments:

ℒinv=TreloarApproximation(): Sets the inverse Langevin approxamation used

Parameters:

μ: Small strain shear modulus

N: Square of the locking stretch of the network.

James HM, Guth E. Theory of the elastic properties of rubber. The Journal of Chemical Physics. 1943 Oct;11(10):455-81.

Hyperelastics.Tube — Method

Tube()

Model:

\[W = \sum\limits_{i=1}^{3}\frac{G_c}{2}(\lambda_i^2-1)+\frac{2Ge}{\beta^2}(\lambda_i^{-\beta}-1)\]

Parameters:

Gc

Ge

β

Heinrich G, Kaliske M. Theoretical and numerical formulation of a molecular based constitutive tube-model of rubber elasticity. Computational and Theoretical Polymer Science. 1997 Jan 1;7(3-4):227-41.

Hyperelastics.ValanisLandel — Method

ValanisLandel()

Model:

\[W = 2\mu\sum\limits_{1}^{3}(\lambda_i(\log\lambda_i -1))\]

Parameters:

μ

Valanis KC, Landel RF. The strain‐energy function of a hyperelastic material in terms of the extension ratios. Journal of Applied Physics. 1967 Jun;38(7):2997-3002.

Hyperelastics.VanDerWaals — Method

VanDerWaals()
VanDerWaals(type)

Model:

\[W = -\mu\{(\lambda_m^2-3)\log(1-\Theta)+\Theta\}-\frac{2\alpha}{3}\bigg(\frac{I-3}{2}\bigg)^{3/2}\]

where:

\[\Theta = \frac{\beta I_1 + (1-\beta)I_2-3}{\lambda_m^2-3)}\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

μ

λm

β

α

Kilian HG, Enderle HF, Unseld K. The use of the van der Waals model to elucidate universal aspects of structure-property relationships in simply extended dry and swollen rubbers. Colloid and Polymer Science. 1986 Oct;264(10):866-76. Ambacher H, Enderle HF, Kilian HG, Sauter A. Relaxation in permanent networks. InRelaxation in Polymers 1989 (pp. 209-220). Steinkopff. Kilian HG. A molecular interpretation of the parameters of the van der Waals equation of state for real networks. Polymer Bulletin. 1980 Sep;3(3):151-8.

Hyperelastics.VerondaWestmann — Method

VerondaWestmann()
VerondaWestmann(type)

Model:

\[W = C_1 (\exp(\alpha(I_1 - 3)) - 1) + C_2 (I_2 - 3)\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

C1

C2

α

Veronda DR, Westmann RA. Mechanical characterization of skin—finite deformations. Journal of biomechanics. 1970 Jan 1;3(1):111-24.

Hyperelastics.Vito — Method

Vito()
Vito(type)

Model:

\[W = \alpha (\exp\bigg(\beta (I_1 - 3)\bigg) + \gamma (I_2 - 3)) - 1)\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

α

β

γ

Vito R. A note on arterial elasticity. Journal of Biomechanics. 1973 Sep 1;6(5):561-4.

Hyperelastics.YamashitaKawabata — Method

YamashitaKawabata()
YamashitaKawabata(type)

Model:

\[W = C_1(I_1-3)+C_2(I_2-3)+\frac{C_3}{N+1}(I_1-3)^{N+1}\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

C1

C2

C3

N

Yamashita Y, Kawabata S. Approximated form of the strain energy-density function of carbon-black filled rubbers for industrial applications. Nippon Gomu Kyokaishi(Journal of the Society of Rubber Industry, Japan)(Japan). 1992;65(9):517-28.

Hyperelastics.Yeoh — Method

Yeoh()
Yeoh(type)

Model:

\[W = \sum\limits_{i,j=0}^{3, 0}C_{i,j}(I_1-3)^i(I_2-3)^j\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or InvariantForm()

Parameters:

C10

C20

C30

Yeoh OH. Characterization of elastic properties of carbon-black-filled rubber vulcanizates. Rubber chemistry and technology. 1990 Nov;63(5):792-805.

Hyperelastics.YeohFleming — Method

YeohFleming()
YeohFleming(type)

Model:

\[W = \frac{A}{B}(1-\exp{-B(I_1-3)}) - C_{10}(I_m-3)\log{1-\frac{I_1-3}{I_m-3}}\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or `InvariantForm()

Parameters:

A

B

C10

Im

Yeoh OH, Fleming PD. A new attempt to reconcile the statistical and phenomenological theories of rubber elasticity. Journal of Polymer Science Part B: Polymer Physics. 1997 Sep 15;35(12):1919-31.

Hyperelastics.Zhao — Method

Zhao()
Zhao(type)

Model:

\[W = C_{-1}^{1}*(I_2-3)+C_{1}^{1}(I_1-3)+C_{2}^{1}(I_1^2-2I_2-3)+C_{2}^{2}(I_1^2-2I_2-3)^2\]

Arguments:

type=PrincipalValueForm(): Sets the form of the strain energy density function. Either PrincipalValueForm() or InvariantForm()

Parameters:

C₋₁¹

C₁¹

C₂¹

C₂²

Zhao Z, Mu X, Du F. Modeling and verification of a new hyperelastic modelfor rubber-like materials. Mathematical Problems in Engineering. 2019 May 22019.

Hyperelastics.ZunigaBeatty — Method

ZunigaBeatty(; ℒinv)

Model:

\[W = \sqrt{\frac{N_3+N_8}{2N_3}}W_{3Chain}+\sqrt{\frac{I_1}{3N_8}}W_{8Chain}\]

Arguments:

ℒinv=TreloarApproximation(): Sets the inverse Langevin approxamation used

Parameters:

μ

N₃

N₈

Elı́as-Zúñiga A, Beatty MF. Constitutive equations for amended non-Gaussian network models of rubber elasticity. International journal of engineering science. 2002 Dec 1;40(20):2265-94.

Compressible Models

Data Driven Models

Hyperelastics.DataDrivenAverageChainBehavior — Method

DataDrivenAverageChainBehavior(data; fchain)

Model:

Adapted from the code provided in the article's supplementary material

Parameters:

None

Fields:

data: A Uniaxial Hyperelastic Test

fchain(λch, pch): A constructor for an approximation with the form f(x, y) => f̂(x) = y

Amores VJ, Benítez JM, Montáns FJ. Average-chain behavior of isotropic incompressible polymers obtained from macroscopic experimental data. A simple structure-based WYPiWYG model in Julia language. Advances in Engineering Software. 2019 Apr 1;130:41-57.

Amores VJ, Benítez JM, Montáns FJ. Data-driven, structure-based hyperelastic manifolds: A macro-micro-macro approach to reverse-engineer the chain behavior and perform efficient simulations of polymers. Computers & Structures. 2020 Apr 15;231:106209.

Hyperelastics.SussmanBathe — Method

SussmanBathe(data; interpolant, k)

Model:

See paper

Parameters:

None

Fields:

data: Hyperelastic Uniaxial test to be used for determining the interpolation

k: Order of the summation in the model.

interpolant: Function of the form, f(s, λ) which returns a function f(λ) = s

Sussman T, Bathe KJ. A model of incompressible isotropic hyperelastic material behavior using spline interpolations of tension–compression test data. Communications in numerical methods in engineering. 2009 Jan;25(1):53-63.

Functions

ContinuumMechanicsBase.CauchyStressTensor — Method

CauchyStressTensor(
    ψ::Hyperelastics.AbstractHyperelasticModel{T<:PrincipalValueForm},
    λ⃗::Array{R, 1},
    p;
    kwargs...
) -> Any

Returns the Cauchy stress tensor for the hyperelastic model ψ with the principle stretches λ⃗ with parameters p.

Arguments:

ψ::AbstractHyperelasticModel: Hyperelastic model

λ⃗::Vector: Vector of principal stretches

p: Model parameters

ad_type: Automatic differentiation backend (see ADTypes.jl)

ContinuumMechanicsBase.CauchyStressTensor — Method

CauchyStressTensor(
    ψ::Hyperelastics.AbstractHyperelasticModel{T<:PrincipalValueForm},
    F::Array{S, 2},
    p;
    kwargs...
)

Returns the Cauchy stress tensor for the hyperelastic model ψ with the deformation gradient F with parameters p.

Arguments:

ψ::AbstractHyperelasticModel: Hyperelastic model

F::Matrix: Deformation gradient matrix

p: Model parameters

ad_type: Automatic differentiation backend (see ADTypes.jl)

ContinuumMechanicsBase.I₁ — Method

I₁(λ⃗)

First stretch invariant - Currently requires the addition of 5 times the machine precision to allow AD to work correctly

$I_1(\vec{\lambda}) = \lambda_1^2+\lambda_2^2+\lambda_3^2 + 5\varepsilon$

ContinuumMechanicsBase.I₂ — Method

I₂(λ⃗)

Second Stretch invariant

$I_2(\vec{\lambda}) = \lambda_1^{-2}+\lambda_2^{-2}+\lambda_3^{-2}$

ContinuumMechanicsBase.I₃ — Method

I₃(λ⃗)

Third Stretch invariant

$I_3(\vec{\lambda}) = (\lambda_1\lambda_2\lambda_3)^2$

ContinuumMechanicsBase.J — Method

J(λ⃗)

Volumetric Stretch

$J(\vec{\lambda}) = \lambda_1\lambda_2\lambda_3$

ContinuumMechanicsBase.SecondPiolaKirchoffStressTensor — Method

SecondPiolaKirchoffStressTensor(
    ψ::Hyperelastics.AbstractHyperelasticModel{T<:PrincipalValueForm},
    F::Array{R, 2},
    p;
    kwargs...
) -> Any

Returns the second PK stress tensor for the hyperelastic model ψ with the deformation gradient F with parameters p.

Arguments:

ψ::AbstractHyperelasticModel: Hyperelastic model

F::Matrix: Deformation gradient matrix

p: Model parameters

ad_type: Automatic differentiation backend (see ADTypes.jl)

ContinuumMechanicsBase.SecondPiolaKirchoffStressTensor — Method

SecondPiolaKirchoffStressTensor(
    ψ::Hyperelastics.AbstractHyperelasticModel{T<:PrincipalValueForm},
    λ⃗::Array{R, 1},
    p;
    ad_type,
    kwargs...
) -> Any

Returns the second PK stress tensor for the hyperelastic model ψ with the principle stretches λ⃗ with parameters p.

Arguments:

ψ::AbstractHyperelasticModel: Hyperelastic model

λ⃗::Vector: Vector of principal stretches

p: Model parameters

ad_type: Automatic differentiation backend (see ADTypes.jl)

ContinuumMechanicsBase.StrainEnergyDensity — Method

StrainEnergyDensity(
    ψ::Hyperelastics.AbstractHyperelasticModel{T<:PrincipalValueForm},
    F::Array{R, 2},
    p
) -> Any

Returns a function for the strain energy density function for the hyperelastic model based on calculating the principal stretches of the deformation gradient, F. The eigen values are found by the following procedure:

\[C = F^T \cdot F a = transpose(eigvecs(C)) C^\ast = (U^\ast)^2 = a^T \cdot C \cdot a \vec{\lambda} = diag(U)\]

Arguments:

ψ::AbstractHyperelasticModel: Hyperelastic model

F::Matrix: Deformation gradient matrix

p: Model parameters

ContinuumMechanicsBase.StrainEnergyDensity — Method

StrainEnergyDensity(
    ψ::Hyperelastics.AbstractHyperelasticModel{T},
    _::Array{R, 1},
    p
) -> Any

Returns the strain energy density for the hyperelastic model ψ with the principle stretches λ⃗ with parameters p.

Arguments:

ψ::AbstractHyperelasticModel: Hyperelastic model

λ⃗::Vector: Vector of principal stretches

p: Model parameters

Hyperelastics.HyperelasticProblem — Function

Creates an OptimizationProblem for use in Optimization.jl to find the optimal parameters.

Arguments:

ψ: material model to use

test or tests: A single or vector of hyperelastics tests to use when fitting the parameters

u₀: Initial guess for parameters

ps: Any additional parameters for calling predict

adb: Select differentiation type from ADTypes.jl. The type is automatically applied to the type of AD applied to the Optimization Problem also.

loss: Loss function from LossFunctions.jl

Hyperelastics.Kawabata1981 — Method

Kawabata1981(λ₁)

Biaxial experimental data from Kawabata et al. The data is more challenging to correctly fit a hyperelastic model to and is proposed as a better test than the Treloar1944 simple tension dataset. Available data is for fixed λ₁ of:

1.040

1.060

1.080

1.100

1.120

1.14

1.16

1.2

1.24

1.3

1.6

1.9

2.2

2.5

2.8

3.1

3.4

3.7

Arguments:

λ₁::Float64: Specification of λ₁ stretch for the data.

Kawabata S, Matsuda M, Tei K, Kawai H. Experimental survey of the strain energy density function of isoprene rubber vulcanizate. Macromolecules. 1981 Jan;14(1):154-62.

Hyperelastics.Treloar1944Uniaxial — Method

Treloar1944Uniaxial()

Uniaxial data for tension of 8% S Rubber at 20C from Fig 3 of Treloar 1944. This is commonly used for testing hyperelastic models.

Treloar LR. Stress-strain data for vulcanized rubber under various types of deformation. Rubber Chemistry and Technology. 1944 Dec;17(4):813-25.

Hyperelastics.parameter_bounds — Method

parameter_bounds(
    _::Hyperelastics.AbstractHyperelasticModel,
    _::Hyperelastics.AbstractHyperelasticTest
) -> NamedTuple{(:lb, :ub), <:Tuple{NamedTuple{(:Gc, :N), <:Tuple{Float64, Any}}, Nothing}}

Returns a tuple of the parameter bounds provided the experimental data and model

Arguments:

ψ::AbstractHyperelasticModel: Hyperelastic model

test or tests: The test or vector of tests to use in finding the parameter bounds.

Hyperelastics.parameters — Method

parameters(
    ψ::Hyperelastics.AbstractHyperelasticModel
) -> NTuple{4, Symbol}

Returns a tuple of the parameters required for the model

Arguments:

ψ::AbstractHyperelasticModel: Hyperelastics model